refactor: 에러 처리 로직 위치 이동 및 메세지 상수화
