+++ __tests__/validation.test.js
@@ -1,5 +1,5 @@
 import { validateSearchKeyword } from '../src/js/Manager/validation';
-import { ERROR_MESSAGE } from '../src/js/Manager/constants';
+import { ERROR_MESSAGE } from '../src/js/constants';
 
 describe('유효성 검사 테스트', () => {
   it('검색 키워드는 2자 이상이어야 한다', () => {
